+++ web-frontend/modules/core/utils/auth.js
@@ -2,7 +2,10 @@ const cookieTokenName = 'jwt_token'
 
 export const setToken = (token, cookie) => {
   if (process.SERVER_BUILD) return
-  cookie.set(cookieTokenName, token)
+  cookie.set(cookieTokenName, token, {
+    path: '/',
+    maxAge: 60 * 60 * 24 * 7,
+  })
 }
 
 export const unsetToken = (cookie) => {